+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v1.5.0 [(2019-6-13)](https://github.com/CalvinVon/axios-api-module/compare/v1.4.1...v1.5.0)
+### Features
+- add `globalPostRequestMiddleWare` static method
+- add `registerPostRequestMiddleWare` instance method
+
 ## v1.4.1 [(2019-5-31)](https://github.com/CalvinVon/axios-api-module/compare/v1.4.0...v1.4.1)
 ### Bug Fixes
 - fix the bug that the `data` option would be *parsed* first before processing in `foreRequestHook` function.